Add docker file and docker compose

+++ src/user/user.service.ts
@@ -4,11 +4,11 @@ import {
   NotFoundException,
 } from '@nestjs/common'
 import { InjectRepository } from '@nestjs/typeorm'
-import { UserEntity } from 'src/domain/UserEntity'
-import { CreateUserDto, EditUserDto } from 'src/model/CreateUserDto'
 import { Repository } from 'typeorm'
 import { v4 } from 'uuid'
 import * as bcrypt from 'bcrypt'
+import { UserEntity } from 'src/domain/UserEntity'
+import { CreateUserDto, EditUserDto } from 'src/model/CreateUserDto'
 import { TransactionEntity } from 'src/domain/TransactionEntity'
 
 @Injectable()
@@ -24,7 +24,6 @@ export class UserService {
     const id = v4()
     const saltRounds = 10
     const hash = await bcrypt.hash(createDto.password, saltRounds)
-
     try {
       await this.userRepository.insert({
         ...createDto,
@@ -32,7 +31,7 @@ export class UserService {
         password: hash,
       })
     } catch (e) {
-      throw new BadRequestException('username is redundant')
+      throw new BadRequestException('username should be unique')
     }
     return id
   }
@@ -42,6 +41,9 @@ export class UserService {
       where: { id },
       relations: { transactions: true, profits: true },
     })
+    if (!user) {
+      throw new NotFoundException('user not found')
+    }
     const transactionDateEntity = await this.transactionRepository.findOne({
       where: { userId: id },
       select: { date: true },
@@ -58,6 +60,9 @@ export class UserService {
       where: { username },
       relations: { transactions: true, profits: true },
     })
+    if (!user) {
+      throw new NotFoundException('user not found')
+    }
     const transactionDateEntity = await this.transactionRepository.findOne({
       where: { userId: user.id },
       select: { date: true },
@@ -84,11 +89,11 @@ export class UserService {
 
   async updateUser(id: string, createDto: Partial<EditUserDto>) {
     const user = await this.getUserById(id)
+    if (!user) {
+      throw new NotFoundException('user not found')
+    }
     const saltRounds = 10
     const hash = await bcrypt.hash(createDto.password, saltRounds)
-    if (!user) {
-      throw new NotFoundException()
-    }
     return this.userRepository.update(id, {
       ...createDto,
       password: createDto.password ? hash : undefined,
@@ -97,6 +102,9 @@ export class UserService {
 
   async deleteUser(id: string) {
     const user = await this.getUserById(id)
+    if (!user) {
+      throw new NotFoundException('user not found')
+    }
     this.userRepository.remove(user)
   }
 }
